Move experiment modals to components
