+++ web/src/lib/components/modals/delete-confirmation-modal.svelte
