+++ web/src/lib/components/modals/edit-experiment-modal.svelte
@@ -1,48 +1,87 @@
 <script lang="ts">
-  import type { HyperParam, Experiment, Visibility } from "$lib/types";
   import {
-    Plus,
     X,
-    Tag as TagIcon,
-    Settings,
-    Beaker,
+    Save,
+    TagIcon,
+    Plus,
     Link,
     Globe,
     Lock,
     ChevronDown,
   } from "lucide-svelte";
-  import { onMount, onDestroy } from "svelte";
-  import { closeCreateExperimentModal } from "$lib/state/app.svelte.js";
   import { enhance } from "$app/forms";
-  import { goto } from "$app/navigation";
+  import type { Experiment } from "$lib/types";
+  import { onMount, onDestroy } from "svelte";
+  import { closeEditExperimentModal } from "$lib/state/app.svelte.js";
 
   let {
-    workspace,
-    experiments,
-  }: { workspace: any; experiments: Experiment[] } = $props();
+    experiment = $bindable(),
+    experiments = $bindable(),
+  }: {
+    experiment: Experiment;
+    experiments: Experiment[];
+  } = $props();
 
-  let hyperparams = $state<HyperParam[]>([]);
-  let addingNewTag = $state<boolean>(false);
+  let experimentCopy = $state<Experiment>({
+    id: experiment.id,
+    name: experiment.name,
+    description: experiment.description,
+    visibility: experiment.visibility,
+    tags: experiment.tags ? [...experiment.tags] : [],
+    availableMetrics: experiment.availableMetrics
+      ? [...experiment.availableMetrics]
+      : [],
+    hyperparams: experiment.hyperparams ? [...experiment.hyperparams] : [],
+    createdAt: experiment.createdAt,
+    updatedAt: experiment.updatedAt,
+  });
+
+  let addingNewTag = $state(false);
   let tag = $state<string | null>(null);
-  let tags = $state<string[]>([]);
-  let visibility = $state<Visibility>("PRIVATE");
-
-  function addTag() {
-    if (tag) {
-      tags = [...tags, tag];
-      tag = null;
-      addingNewTag = false;
-    }
-  }
-
   let reference = $state<Experiment | null>(null);
   let searchInput = $state<string>("");
 
   let filteredExperiments = $derived(
-    experiments.filter((exp) =>
-      exp.name.toLowerCase().includes(searchInput.toLowerCase()),
+    experiments.filter(
+      (exp) =>
+        exp.id !== experiment.id &&
+        exp.name.toLowerCase().includes(searchInput.toLowerCase()),
     ),
   );
+
+  onMount(async () => {
+    document.body.classList.add("overflow-hidden");
+
+    try {
+      reference = null;
+      const response = await fetch(`/api/experiments/${experiment.id}/ref`);
+      if (response.ok) {
+        const referenceIds = await response.json();
+        const referencesToLoad = referenceIds.filter(
+          (id: String) => id !== experiment.id,
+        );
+
+        if (referencesToLoad.length > 0) {
+          try {
+            const refResponse = await fetch(
+              `/api/experiments/${referencesToLoad[0]}`,
+            );
+            if (refResponse.ok) {
+              reference = await refResponse.json();
+            }
+          } catch (refError) {
+            console.error("Failed to load referenced experiment:", refError);
+          }
+        }
+      }
+    } catch (error) {
+      console.error("Failed to load references:", error);
+    }
+  });
+
+  onDestroy(() => {
+    document.body.classList.remove("overflow-hidden");
+  });
 
   function selectReference(exp: Experiment) {
     reference = exp;
@@ -56,55 +95,79 @@
     searchInput = "";
   }
 
-  onMount(() => {
-    document.body.classList.add("overflow-hidden");
-  });
-
-  onDestroy(() => {
-    document.body.classList.remove("overflow-hidden");
-  });
+  function addTag(e: KeyboardEvent | MouseEvent) {
+    e.preventDefault();
+    if (tag && tag !== "") {
+      if (!experimentCopy.tags) {
+        experimentCopy.tags = [];
+      }
+      experimentCopy.tags.push(tag);
+      tag = null;
+    }
+  }
 </script>
 
 <div
   class="fixed inset-0 bg-ctp-base/90 backdrop-blur-sm
          flex items-center justify-center p-4 z-50 overflow-hidden font-mono"
 >
+  <!-- MODAL CONTAINER -->
   <div
-    class="w-full max-w-xl bg-ctp-base border border-ctp-surface0/30 overflow-auto max-h-[90vh]"
+    class="bg-ctp-base w-full max-w-xl border border-ctp-surface0/30 overflow-auto max-h-[90vh]"
     role="dialog"
     aria-modal="true"
     aria-labelledby="modal-title"
   >
+    <!-- HEADER -->
     <div
       class="flex items-center justify-between p-4 border-b border-ctp-surface0/20"
     >
       <div class="flex items-center gap-3">
         <div class="w-2 h-6 bg-ctp-mauve rounded-full"></div>
         <div>
-          <h3 id="modal-title" class="text-lg text-ctp-text font-mono">
-            New Experiment
-          </h3>
-          <div class="text-xs text-ctp-subtext0">create experiment config</div>
+          <h2 id="modal-title" class="text-lg text-ctp-text font-mono">
+            Edit Experiment
+          </h2>
+          <div class="text-xs text-ctp-subtext0">modify experiment config</div>
         </div>
       </div>
+      <button
+        onclick={() => {
+          closeEditExperimentModal();
+        }}
+        type="button"
+        class="text-ctp-subtext0 hover:text-ctp-red hover:bg-ctp-surface0/30 rounded p-1 transition-all"
+      >
+        <X size={14} />
+      </button>
     </div>
 
+    <!-- FORM -->
     <form
       method="POST"
-      action="?/create"
+      action="/experiments?/update"
       class="p-4 space-y-4"
-      use:enhance={({ formElement, formData, action, cancel }) => {
-        formData.append("workspace-id", workspace.id);
+      use:enhance={({ formElement, formData, action, cancel, submitter }) => {
         return async ({ result, update }) => {
-          if (result.type === "redirect") {
-            goto(result.location);
-          } else {
-            await update();
-            closeCreateExperimentModal();
+          if (result.type === "success" || result.type === "redirect") {
+            experiment.name = experimentCopy.name;
+            experiment.description = experimentCopy.description;
+            experiment.visibility = experimentCopy.visibility;
+            experiment.tags = [...experiment.tags];
           }
+          closeEditExperimentModal();
+
+          await update();
         };
       }}
     >
+      <input
+        class="hidden"
+        id="experiment-id"
+        name="experiment-id"
+        value={experimentCopy.id}
+      />
+
       <div class="space-y-3">
         <!-- Basic config -->
         <div class="space-y-1 text-xs overflow-hidden">
@@ -112,10 +175,12 @@
             <span class="text-ctp-subtext0">name</span>
             <span class="text-ctp-text">=</span>
             <input
+              id="experiment-name"
               name="experiment-name"
               type="text"
               class="bg-ctp-surface0/20 border border-ctp-surface0/30 px-2 py-1 text-ctp-text placeholder-ctp-subtext0 focus:outline-none focus:ring-1 focus:ring-ctp-mauve focus:border-ctp-mauve transition-all text-xs min-w-0"
               placeholder="experiment_name"
+              bind:value={experimentCopy.name}
               required
             />
           </div>
@@ -123,10 +188,12 @@
             <span class="text-ctp-subtext0">desc</span>
             <span class="text-ctp-text">=</span>
             <textarea
+              id="experiment-description"
               name="experiment-description"
               rows="2"
               class="bg-ctp-surface0/20 border border-ctp-surface0/30 px-2 py-1 text-ctp-text placeholder-ctp-subtext0 focus:outline-none focus:ring-1 focus:ring-ctp-mauve focus:border-ctp-mauve transition-all resize-none text-xs min-w-0"
               placeholder="description"
+              bind:value={experimentCopy.description}
               required
             ></textarea>
           </div>
@@ -135,16 +202,21 @@
         <!-- Visibility Setting -->
         <div class="space-y-1">
           <div class="text-sm text-ctp-text">Visibility</div>
-          <input type="hidden" name="visibility" value={visibility} />
+          <input
+            type="hidden"
+            id="edit-visibility-input"
+            name="visibility"
+            bind:value={experimentCopy.visibility}
+          />
 
           <div class="flex gap-2 text-xs">
             <button
               type="button"
               class={"flex items-center gap-1 px-2 py-1 transition-colors " +
-                (visibility === "PUBLIC"
+                (experimentCopy.visibility === "PUBLIC"
                   ? "bg-ctp-green/20 text-ctp-green border border-ctp-green/30"
                   : "bg-ctp-surface0/20 text-ctp-subtext0 hover:bg-ctp-surface0/30 hover:text-ctp-text border border-ctp-surface0/30")}
-              onclick={() => (visibility = "PUBLIC")}
+              onclick={() => (experimentCopy.visibility = "PUBLIC")}
             >
               <Globe size={10} />
               <span>public</span>
@@ -153,10 +225,11 @@
             <button
               type="button"
               class={"flex items-center gap-1 px-2 py-1 transition-colors " +
-                (visibility === "PRIVATE"
+                (experimentCopy.visibility === "PRIVATE" ||
+                !experimentCopy.visibility
                   ? "bg-ctp-red/20 text-ctp-red border border-ctp-red/30"
                   : "bg-ctp-surface0/20 text-ctp-subtext0 hover:bg-ctp-surface0/30 hover:text-ctp-text border border-ctp-surface0/30")}
-              onclick={() => (visibility = "PRIVATE")}
+              onclick={() => (experimentCopy.visibility = "PRIVATE")}
             >
               <Lock size={10} />
               <span>private</span>
@@ -180,22 +253,24 @@
             </summary>
             <div class="pt-2 pl-6">
               <div class="flex flex-wrap items-center gap-2">
-                {#each tags as tag, i}
-                  <input type="hidden" value={tag} name="tags.{i}" />
-                  <span
-                    class="inline-flex items-center px-2 py-0.5 text-xs bg-ctp-blue/10 text-ctp-blue border-0 group"
-                  >
-                    {tag}
-                    <button
-                      type="button"
-                      class="text-ctp-blue/70 hover:text-ctp-red transition-colors ml-1.5"
-                      onclick={() => tags.splice(i, 1)}
-                      aria-label="Remove tag"
+                {#if experimentCopy.tags && experimentCopy.tags.length > 0}
+                  {#each experimentCopy.tags as tag, i}
+                    <input type="hidden" value={tag} name="tags.{i}" />
+                    <span
+                      class="inline-flex items-center px-2 py-0.5 text-xs bg-ctp-blue/10 text-ctp-blue border-0 group"
                     >
-                      <X size={12} />
-                    </button>
-                  </span>
-                {/each}
+                      {tag}
+                      <button
+                        type="button"
+                        class="text-ctp-blue/70 hover:text-ctp-red transition-colors ml-1.5"
+                        onclick={() => experimentCopy.tags?.splice(i, 1)}
+                        aria-label="Remove tag"
+                      >
+                        <X size={12} />
+                      </button>
+                    </span>
+                  {/each}
+                {/if}
 
                 {#if addingNewTag}
                   <div class="flex items-center gap-1">
@@ -204,19 +279,11 @@
                       bind:value={tag}
                       class="w-40 px-2 py-1.5 text-sm bg-ctp-base border-0 rounded-lg text-ctp-text focus:outline-none focus:ring-2 focus:ring-ctp-blue transition-all placeholder-ctp-overlay0 shadow-sm"
                       placeholder="New tag"
-                      onkeydown={(event) => {
-                        if (event.key === "Enter") {
-                          event.preventDefault();
-                          addTag();
-                        }
-                      }}
+                      onkeydown={addTag}
                     />
                     <button
                       type="button"
-                      onclick={(event) => {
-                        event.preventDefault();
-                        addTag();
-                      }}
+                      onclick={addTag}
                       class="p-1.5 rounded-full text-ctp-blue hover:bg-ctp-blue/10 transition-all"
                     >
                       <Plus size={14} />
@@ -225,8 +292,8 @@
                 {:else}
                   <button
                     type="button"
-                    onclick={(event) => {
-                      event.preventDefault();
+                    onclick={(e) => {
+                      e.preventDefault();
                       addingNewTag = true;
                     }}
                     class="inline-flex items-center gap-1 py-0.5 px-2 text-xs bg-transparent text-ctp-blue border border-dashed border-ctp-blue/50 hover:bg-ctp-blue/10 transition-all"
@@ -235,57 +302,6 @@
                     Add Tag
                   </button>
                 {/if}
-              </div>
-            </div>
-          </details>
-
-          <!-- Parameters Section -->
-          <details class="group">
-            <summary
-              class="flex items-center gap-2 cursor-pointer text-ctp-subtext0 hover:text-ctp-text py-1.5"
-            >
-              <Settings size={16} class="text-ctp-sapphire" />
-              <span class="text-sm">parameters</span>
-              <ChevronDown
-                size={16}
-                class="ml-auto text-ctp-subtext0 group-open:rotate180"
-              />
-            </summary>
-            <div class="pt-2 pl-6">
-              <div class="space-y-3">
-                {#each hyperparams as pair, i}
-                  <div class="flex gap-2 items-center">
-                    <input
-                      class="w-full px-3 py-1.5 text-sm bg-ctp-base border-0 rounded-lg text-ctp-text focus:outline-none focus:ring-2 focus:ring-ctp-sapphire transition-all flex-1 placeholder-ctp-overlay0 shadow-sm"
-                      name="hyperparams.{i}.key"
-                      placeholder="Parameter name"
-                      required
-                    />
-                    <input
-                      class="w-full px-3 py-1.5 text-sm bg-ctp-base border-0 rounded-lg text-ctp-text focus:outline-none focus:ring-2 focus:ring-ctp-sapphire transition-all flex-1 placeholder-ctp-overlay0 shadow-sm"
-                      name="hyperparams.{i}.value"
-                      placeholder="Value"
-                      required
-                    />
-                    <button
-                      type="button"
-                      class="p-1.5 text-ctp-subtext0 hover:text-ctp-red hover:bg-ctp-red/10 rounded-full transition-all"
-                      onclick={() => hyperparams.splice(i, 1)}
-                    >
-                      <X size={16} />
-                    </button>
-                  </div>
-                {/each}
-
-                <button
-                  type="button"
-                  class="inline-flex items-center gap-1.5 py-1.5 px-3 text-sm rounded-lg bg-ctp-sapphire/10 text-ctp-sapphire border border-dashed border-ctp-sapphire/50 hover:bg-ctp-sapphire/20 transition-all"
-                  onclick={() =>
-                    (hyperparams = [...hyperparams, { key: "", value: "" }])}
-                >
-                  <Plus size={14} />
-                  Add Parameter
-                </button>
               </div>
             </div>
           </details>
@@ -385,13 +401,13 @@
                       </button>
                     {/each}
 
-                    {#if filteredExperiments.length === 0 && experiments.length > 0}
+                    {#if filteredExperiments.length === 0 && experiments.length > 1}
                       <div class="p-2 text-sm text-ctp-subtext0 text-center">
                         No experiments found
                       </div>
-                    {:else if experiments.length === 0}
+                    {:else if experiments.length <= 1}
                       <div class="p-2 text-sm text-ctp-subtext0 text-center">
-                        No experiments in this workspace yet
+                        No other experiments in this workspace
                       </div>
                     {/if}
                   </div>
@@ -402,13 +418,13 @@
         </div>
       </div>
 
-      <!-- Action Buttons -->
+      <!-- Footer -->
       <div
         class="flex justify-end gap-2 pt-3 mt-3 border-t border-ctp-surface0/20"
       >
         <button
           onclick={() => {
-            closeCreateExperimentModal();
+            closeEditExperimentModal();
           }}
           type="button"
           class="bg-ctp-surface0/20 border border-ctp-surface0/30 text-ctp-subtext0 hover:bg-ctp-surface0/30 hover:text-ctp-text px-3 py-2 text-xs transition-all"
@@ -419,7 +435,7 @@
           type="submit"
           class="bg-ctp-surface0/20 border border-ctp-surface0/30 text-ctp-mauve hover:bg-ctp-mauve/10 hover:border-ctp-mauve/30 px-3 py-2 text-xs transition-all"
         >
-          create
+          update
         </button>
       </div>
     </form>
